+++ src/egon/data/airflow/dags/pipeline.py
@@ -3,6 +3,7 @@ import os
 from airflow.utils.dates import days_ago
 import airflow
 
+from egon.data.config import settings as egon_settings
 from egon.data.config import set_numexpr_threads
 from egon.data.metadata import Json_Metadata
 from egon.data.datasets import database
@@ -105,8 +106,11 @@ from egon.data.datasets.scenario_path import CreateIntermediateScenarios
 # Set number of threads used by numpy and pandas
 set_numexpr_threads()
 
+prefix = egon_settings()["egon-data"].get("--prefix")
+prefix = "" if prefix is None else f"{prefix}-"
+
 with airflow.DAG(
-    "egon-data-processing-pipeline",
+    f"{prefix}egon-data-processing-pipeline",
     description="The eGo^N data processing DAG.",
     default_args={"start_date": days_ago(1)},
     template_searchpath=[